requirements.txt
antlr4-python3-runtime==4.9.3
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
flatbuffers==25.9.23
fonttools==4.58.2
fsspec==2025.5.1
hf-xet==1.2.0
huggingface-hub==0.33.0
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.8
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.5
numpy==2.3.0
omegaconf==2.3.0
onnx==1.19.1
onnxruntime==1.23.2
onnxsim==0.4.36
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.3
pillow==11.2.1
protobuf==6.33.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rich==14.2.0
safetensors==0.5.3
seaborn==0.13.2
segment-anything==1.0
segmentation_models_pytorch==0.5.0
six==1.17.0
sympy==1.14.0
timm==1.0.15
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tqdm==4.67.1
triton==3.5.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
